+++ package.json
@@ -31,9 +31,7 @@
   ],
   "license": "MIT",
   "repository": "kemitchell/spdx-satisfies.js",
-  "files": [
-    "index.js"
-  ],
+  "main": "src/index.js",
   "scripts": {
     "test": "jest",
     "coverage": "jest --coverage",
